refactor: (sample) Change '_' to '-' in filenames for consistency
