+++ js-miniapp-sample/src/pages/uuid-sdk.js
