+++ js-miniapp-sample/src/pages/web-location.js
