+++ js-miniapp-sample/src/pages/window-actions.js
